requirements.txt.txt
pandas==2.1.1
requests==2.31.0
psycopg2==2.9.6
apache-airflow==2.5.0